fix fast-element circular dependencies
